Added Limbo.Umbraco.Video dependency
Besides the dependency. this commit also contains a number of related changes, as well as a breaking change in the way the video data is saved in Umbraco, requiring existing videos to be inserted.

Due to the way Umbraco/JSON.net handles date like string values (eg. string values containing an ISO 8601 date), we need to keep a serialized version the video data received from the YouTube API in order to prevent corrupt timestamps. See this issue for further information: https://github.com/umbraco/Umbraco-CMS/issues/3075