Setting a max width for the property editor
Some of the build-in property editors in Umbraco have a max width of 800px but IMO this is a bit to narrow for some property editors. With the YouTube property editor, I think 1100 px is a better max width. This allows the property editor to have a bit more width on larger screens, while still ensuring that the property editor doesn't take up of full width on ultra wide monitors.