Rewrote internal logic for looking up YouTube videos
Previously a lot of the logic used for looking up videos has been located in the controller, which usually isn't the best approach, as the logic then can be called from elsewhere. With this commit, the "YouTubeService" class now has a "GetIntermediaryVideoValue", which returns an instance of "YouTubeIntermediaryVideoValue" if successful, or throws an "YouTubeException" exception if it fails. More specific exception classes (extending the "YouTubeException" class) may be thrown to better indicate the error.

Among other things, the new "GetIntermediaryVideoValue" method can be used to get the correct value to save in Umbraco when creating or updating content programmatically, as the returned value will follow the same format as saved by the property editor

The value saved by the property editor (and now also represented by the "GetIntermediaryVideoValue" class) is an object with various information, such as GUID key referencing the used YouTube credentials, optinally a list of parameters if used in the URL and embed code, but then also a "video" property encasulating the JSON video object received from the YouTube API. With this commi, the property is now renamed to "details" to align with our other video packages. Both the Angular property editor and the C# property value converter has been updated to support both "video" and "details" to avoid breaking already saved values.